requirements.txt
python-dotenv==0.14.0
pytube==9.7.0
typing-extensions==3.7.4.3
